@@ -1,6 +1,6 @@
 module github.com/MarioCarrion/grpc-microservice-example
 
-go 1.21
+go 1.22
 
 require (
 	buf.build/gen/go/bufbuild/protovalidate/protocolbuffers/go v1.34.2-20240508200655-46a4cf4ba109.2
