@@ -1,6 +1,7 @@
 module github.com/MarioCarrion/grpc-microservice-example/internal/tools
 
-go 1.21
+go 1.22
+
 toolchain go1.22.5
 
 require (
