@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: user/v1/user.proto
 
@@ -269,7 +269,7 @@ func file_user_v1_user_proto_rawDescGZIP() []byte {
 
 var file_user_v1_user_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_user_v1_user_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_user_v1_user_proto_goTypes = []interface{}{
+var file_user_v1_user_proto_goTypes = []any{
 	(MaritalStatus)(0), // 0: user.v1.MaritalStatus
 	(*User)(nil),       // 1: user.v1.User
 	(*Address)(nil),    // 2: user.v1.Address
@@ -290,7 +290,7 @@ func file_user_v1_user_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_user_v1_user_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_user_v1_user_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*User); i {
 			case 0:
 				return &v.state
@@ -302,7 +302,7 @@ func file_user_v1_user_proto_init() {
 				return nil
 			}
 		}
-		file_user_v1_user_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_user_v1_user_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Address); i {
 			case 0:
 				return &v.state
@@ -315,7 +315,7 @@ func file_user_v1_user_proto_init() {
 			}
 		}
 	}
-	file_user_v1_user_proto_msgTypes[0].OneofWrappers = []interface{}{}
+	file_user_v1_user_proto_msgTypes[0].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
