@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: user/v1/user_service.proto
 
@@ -154,7 +154,7 @@ func file_user_v1_user_service_proto_rawDescGZIP() []byte {
 }
 
 var file_user_v1_user_service_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_user_v1_user_service_proto_goTypes = []interface{}{
+var file_user_v1_user_service_proto_goTypes = []any{
 	(*GetUserRequest)(nil),  // 0: user.v1.GetUserRequest
 	(*GetUserResponse)(nil), // 1: user.v1.GetUserResponse
 	(*User)(nil),            // 2: user.v1.User
@@ -177,7 +177,7 @@ func file_user_v1_user_service_proto_init() {
 	}
 	file_user_v1_user_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_user_v1_user_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_user_v1_user_service_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*GetUserRequest); i {
 			case 0:
 				return &v.state
@@ -189,7 +189,7 @@ func file_user_v1_user_service_proto_init() {
 				return nil
 			}
 		}
-		file_user_v1_user_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_user_v1_user_service_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*GetUserResponse); i {
 			case 0:
 				return &v.state
