@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: wearable/v1/wearable_service.proto
 
@@ -422,7 +422,7 @@ func file_wearable_v1_wearable_service_proto_rawDescGZIP() []byte {
 }
 
 var file_wearable_v1_wearable_service_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_wearable_v1_wearable_service_proto_goTypes = []interface{}{
+var file_wearable_v1_wearable_service_proto_goTypes = []any{
 	(*BeatsPerMinuteRequest)(nil),           // 0: wearable.v1.BeatsPerMinuteRequest
 	(*BeatsPerMinuteResponse)(nil),          // 1: wearable.v1.BeatsPerMinuteResponse
 	(*ConsumeBeatsPerMinuteRequest)(nil),    // 2: wearable.v1.ConsumeBeatsPerMinuteRequest
@@ -450,7 +450,7 @@ func file_wearable_v1_wearable_service_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_wearable_v1_wearable_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_wearable_v1_wearable_service_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*BeatsPerMinuteRequest); i {
 			case 0:
 				return &v.state
@@ -462,7 +462,7 @@ func file_wearable_v1_wearable_service_proto_init() {
 				return nil
 			}
 		}
-		file_wearable_v1_wearable_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_wearable_v1_wearable_service_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*BeatsPerMinuteResponse); i {
 			case 0:
 				return &v.state
@@ -474,7 +474,7 @@ func file_wearable_v1_wearable_service_proto_init() {
 				return nil
 			}
 		}
-		file_wearable_v1_wearable_service_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_wearable_v1_wearable_service_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ConsumeBeatsPerMinuteRequest); i {
 			case 0:
 				return &v.state
@@ -486,7 +486,7 @@ func file_wearable_v1_wearable_service_proto_init() {
 				return nil
 			}
 		}
-		file_wearable_v1_wearable_service_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_wearable_v1_wearable_service_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ConsumeBeatsPerMinuteResponse); i {
 			case 0:
 				return &v.state
@@ -498,7 +498,7 @@ func file_wearable_v1_wearable_service_proto_init() {
 				return nil
 			}
 		}
-		file_wearable_v1_wearable_service_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_wearable_v1_wearable_service_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*CalculateBeatsPerMinuteRequest); i {
 			case 0:
 				return &v.state
@@ -510,7 +510,7 @@ func file_wearable_v1_wearable_service_proto_init() {
 				return nil
 			}
 		}
-		file_wearable_v1_wearable_service_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_wearable_v1_wearable_service_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*CalculateBeatsPerMinuteResponse); i {
 			case 0:
 				return &v.state
