@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.4.0
 // - protoc             (unknown)
 // source: wearable/v1/wearable_service.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.62.0 or later.
+const _ = grpc.SupportPackageIsVersion8
 
 const (
 	WearableService_BeatsPerMinute_FullMethodName          = "/wearable.v1.WearableService/BeatsPerMinute"
@@ -42,11 +42,12 @@ func NewWearableServiceClient(cc grpc.ClientConnInterface) WearableServiceClient
 }
 
 func (c *wearableServiceClient) BeatsPerMinute(ctx context.Context, in *BeatsPerMinuteRequest, opts ...grpc.CallOption) (WearableService_BeatsPerMinuteClient, error) {
-	stream, err := c.cc.NewStream(ctx, &WearableService_ServiceDesc.Streams[0], WearableService_BeatsPerMinute_FullMethodName, opts...)
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &WearableService_ServiceDesc.Streams[0], WearableService_BeatsPerMinute_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &wearableServiceBeatsPerMinuteClient{stream}
+	x := &wearableServiceBeatsPerMinuteClient{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -74,11 +75,12 @@ func (x *wearableServiceBeatsPerMinuteClient) Recv() (*BeatsPerMinuteResponse, e
 }
 
 func (c *wearableServiceClient) ConsumeBeatsPerMinute(ctx context.Context, opts ...grpc.CallOption) (WearableService_ConsumeBeatsPerMinuteClient, error) {
-	stream, err := c.cc.NewStream(ctx, &WearableService_ServiceDesc.Streams[1], WearableService_ConsumeBeatsPerMinute_FullMethodName, opts...)
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &WearableService_ServiceDesc.Streams[1], WearableService_ConsumeBeatsPerMinute_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &wearableServiceConsumeBeatsPerMinuteClient{stream}
+	x := &wearableServiceConsumeBeatsPerMinuteClient{ClientStream: stream}
 	return x, nil
 }
 
@@ -108,11 +110,12 @@ func (x *wearableServiceConsumeBeatsPerMinuteClient) CloseAndRecv() (*ConsumeBea
 }
 
 func (c *wearableServiceClient) CalculateBeatsPerMinute(ctx context.Context, opts ...grpc.CallOption) (WearableService_CalculateBeatsPerMinuteClient, error) {
-	stream, err := c.cc.NewStream(ctx, &WearableService_ServiceDesc.Streams[2], WearableService_CalculateBeatsPerMinute_FullMethodName, opts...)
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &WearableService_ServiceDesc.Streams[2], WearableService_CalculateBeatsPerMinute_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &wearableServiceCalculateBeatsPerMinuteClient{stream}
+	x := &wearableServiceCalculateBeatsPerMinuteClient{ClientStream: stream}
 	return x, nil
 }
 
@@ -179,7 +182,7 @@ func _WearableService_BeatsPerMinute_Handler(srv interface{}, stream grpc.Server
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(WearableServiceServer).BeatsPerMinute(m, &wearableServiceBeatsPerMinuteServer{stream})
+	return srv.(WearableServiceServer).BeatsPerMinute(m, &wearableServiceBeatsPerMinuteServer{ServerStream: stream})
 }
 
 type WearableService_BeatsPerMinuteServer interface {
@@ -196,7 +199,7 @@ func (x *wearableServiceBeatsPerMinuteServer) Send(m *BeatsPerMinuteResponse) er
 }
 
 func _WearableService_ConsumeBeatsPerMinute_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(WearableServiceServer).ConsumeBeatsPerMinute(&wearableServiceConsumeBeatsPerMinuteServer{stream})
+	return srv.(WearableServiceServer).ConsumeBeatsPerMinute(&wearableServiceConsumeBeatsPerMinuteServer{ServerStream: stream})
 }
 
 type WearableService_ConsumeBeatsPerMinuteServer interface {
@@ -222,7 +225,7 @@ func (x *wearableServiceConsumeBeatsPerMinuteServer) Recv() (*ConsumeBeatsPerMin
 }
 
 func _WearableService_CalculateBeatsPerMinute_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(WearableServiceServer).CalculateBeatsPerMinute(&wearableServiceCalculateBeatsPerMinuteServer{stream})
+	return srv.(WearableServiceServer).CalculateBeatsPerMinute(&wearableServiceCalculateBeatsPerMinuteServer{ServerStream: stream})
 }
 
 type WearableService_CalculateBeatsPerMinuteServer interface {
